refactor(controlplane): make server events non-internal
