+++ controlplane/event/event.go
@@ -1,4 +1,4 @@
-package registry
+package event
 
 import (
 	"fmt"
